aprs-bot/requirements.txt
-i https://pypi.org/simple
aprs==7.0.0
aprslib==0.6.47
bitarray==1.2.1
certifi==2019.11.28
chardet==3.0.4
hexdump==3.3
idna==2.9
kiss==7.0.0
paho-mqtt==1.5.0
pyserial==3.4
requests==2.23.0
urllib3==1.25.8
